Add profile UI to display submitted profile data
