+++ src/app/client-dashboard/profile/loading.tsx
@@ -1,0 +1,7 @@
+export default function Loading() {
+  return (
+    <div className="fixed inset-0 flex items-center justify-center ">
+      <div className="animate-spin h-12 w-12 border-t-4 border-primary-orange rounded-full" />
+    </div>
+  );
+}
